refactor: impove code structure
