+++ apps/api/Modules/Inventory/Inventory/Products/Features/AdjustProductStock/AdjustProductStockContracts.cs
@@ -1,0 +1,4 @@
+﻿namespace Inventory.Products.Features.AdjustProductStock;
+
+public record AdjustProductStockRequest(Guid Id, int Delta);
+public record AdjustProductStockResponse(Guid Id);
