+++ apps/api/Modules/Inventory/Inventory/Products/Features/CreateProduct/CreateProductContracts.cs
@@ -1,0 +1,4 @@
+﻿namespace Inventory.Products.Features.CreateProduct;
+
+public record CreateProductRequest(ProductDto Product);
+public record CreateProductResponse(Guid Id);
