+++ apps/api/Modules/Inventory/Inventory/Products/Features/GetLowStockProducts/GetLowStockProductsContracts.cs
@@ -1,0 +1,3 @@
+﻿namespace Inventory.Products.Features.GetLowStockProducts;
+
+public record GetLowStockProductsResponse(IEnumerable<ProductDto> Products);
