+++ apps/api/Modules/Inventory/Inventory/Products/Features/GetProducts/GetProductsContracts.cs
@@ -1,0 +1,3 @@
+﻿namespace Inventory.Products.Features.GetProducts;
+
+public record GetProductsResponse(IEnumerable<ProductDto> Products);
