--- apps/api/Modules/Inventory/Inventory/Products/Features/GetProducts/GetProductsEndpoint.cs
@@ -1,34 +1,0 @@
-﻿using Carter;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Routing;
-using Shared.Messaging;
-
-namespace Inventory.Products.Features.GetProducts;
-
-public record GetProductsResponse(IEnumerable<ProductDto> Products);
-
-public class GetProductsEndpoint : ICarterModule
-{
-    public void AddRoutes(IEndpointRouteBuilder app)
-    {
-        app.MapGet("/products", async (IMediator sender) =>
-        {
-            var result = await sender.Send(new GetProductsQuery());
-
-            var response = new List<ProductDto>();
-            foreach(var product in result.Products)
-            {
-                var productDto = new ProductDto(product.Name, product.Stock, product.Threshold);
-                response.Add(productDto);
-            }
-
-            return Results.Ok(response);
-        })
-        .WithName("GetProducts")
-        .Produces<GetProductsResponse>(StatusCodes.Status200OK)
-        .ProducesProblem(StatusCodes.Status400BadRequest)
-        .WithSummary("Get Products")
-        .WithDescription("Get Products");
-    }
-}
