+++ apps/api/Modules/Inventory/Inventory/Products/Features/SetProductThreshold/SetProductThresholdContracts.cs
@@ -1,0 +1,4 @@
+﻿namespace Inventory.Products.Features.SetProductThreshold;
+
+public record SetProductThresholdRequest(Guid Id, int Threshold);
+public record SetProductThresholdResponse(Guid Id);
